update: added mempool metrics
