+++ crates/client/devnet/src/lib.rs
@@ -190,7 +190,7 @@ mod tests {
     use mc_block_import::{BlockImporter, BlockValidationContext};
     use mc_db::MadaraBackend;
     use mc_mempool::block_production::BlockProductionTask;
-    use mc_mempool::metrics::BlockProductionMetrics;
+    use mc_mempool::block_production_metrics::BlockProductionMetrics;
     use mc_mempool::MempoolProvider;
     use mc_mempool::{transaction_hash, L1DataProvider, Mempool, MockL1DataProvider};
